+++ src/egon/data/airflow/dags/pipeline.py
@@ -459,6 +459,7 @@ with airflow.DAG(
             scenario_capacities,
             Vg250MvGridDistricts,
             chp,
+            scenario_parameters,
         ]
     )