+++ CHANGELOG.rst
@@ -295,7 +295,8 @@ Changed
   `#545 <https://github.com/openego/eGon-data/issues/545>`_
 * Integrate fuel and CO2 costs for eGon2035 to scenario parameters
   `#549 <https://github.com/openego/eGon-data/issues/549>`_
-
+* Add control parameter to scenario parameter
+  `#572 <https://github.com/openego/eGon-data/issues/572>`_
 
 Bug fixes
 ---------